// src/type/type.ts
export type userType = {
    avatar_url?: string;
    name?: string;
    login?: string;
    created_at?: string;
    bio?: string;
    public_repos?: number;
    followers?: number;
    following?: number;
    location?: string;
    blog?: string;
    company?: string;
    twitter_username?:string;
}